requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.24
plotly>=5.18
requests>=2.31
scikit-learn>=1.3
networkx>=3.1
python-dotenv>=1.0
